pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "parenting-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "langchain-text-splitters>=1.1.1",
    "langsmith>=0.7.22",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.27.2.2",
    "pymupdf4llm>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
    "uvicorn>=0.42.0",
]

[project.scripts]
app-dev = "src.api.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uvicorn]
app = "src.api.main:app"
host = "0.0.0.0"
port = 8000
reload = true
